chiluServer/requirements.txt
asgiref==3.2.10
chardet==3.0.4
dj-database-url==0.5.0
Django==3.1.1
gunicorn==20.0.4
numpy==1.19.4
opencv-python==4.4.0.46
Pillow==8.0.1
scikit-learn==0.23.2
scipy==1.5.4
sqlparse==0.3.1

